requirements.txt
anthropic>=0.40.0
openai>=1.0.0
praw>=7.7.0
google-api-python-client>=2.100.0
feedparser>=6.0.0
tweepy>=4.14.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
schedule>=1.2.0
python-telegram-bot>=21.0
